Move notify work to its own blueprint
