+++ app/templates/notify-settings/notify-settings--reply-to.html
@@ -4,7 +4,7 @@
 <div class="grid-row">
   <div class="grid-column-two-thirds" id="options">
 
-    <form method=post enctype=multipart/form-data action="{{ url_for("main.notify_settings_api_key") }}">
+    <form method=post enctype=multipart/form-data action="{{ url_for("notify.notify_settings_api_key") }}">
       {% if error_message %}
       <blockquote class="error warning-text">
         <p><strong>An error has occurred</strong></p>
