+++ app/templates/notify-settings/notify-settings--template-ids.html
@@ -4,7 +4,7 @@
 <div class="grid-row">
   <div class="grid-column-two-thirds" id="options">
 
-    <form method=post enctype=multipart/form-data action="{{ url_for("main.notify_settings_reply_id") }}">
+    <form method=post enctype=multipart/form-data action="{{ url_for("notify.notify_settings_reply_id") }}">
       {% if error_message %}
       <blockquote class="error warning-text">
         <p><strong>An error has occurred</strong></p>
