+++ app/templates/notify-settings/notify-settings-intro.html
@@ -13,7 +13,7 @@
             Notify service and upload your spreadsheet there</p>
     <hr>
     <div class="button-group">
-      <a href="{{ url_for("main.notify_settings_template_id") }}" class="button button--action">Start now</a>
+      <a href="{{ url_for("notify.notify_settings_template_id") }}" class="button button--action">Start now</a>
     </div>
     </div>
   </div>
